Lista för användare, in/utloggningsfunktion samt fixat att testerna hamnat i testmappen
