+++ todogrupp5/pom.xml
@@ -14,7 +14,7 @@
 	<name>todogrupp5</name>
 	<description>Demo project for Spring Boot</description>
 	<properties>
-		<java.version>21</java.version>
+		<java.version>17</java.version>
 	</properties>
 	<dependencies>
 		<dependency>
