+++ todogrupp5/src/main/java/com/todogrupp5/todogrupp5/UserList.java
@@ -31,6 +31,4 @@ public class UserList {
     public void setUsers(List<User> users) {
         this.users = users;
     }
-
-    
 }
