+++ todogrupp5/src/main/resources/templates/index.html
@@ -1,12 +1,19 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="en" xmlns="http://www.thymeleaf.org">
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-    <link rel="stylesheet" href="style.css">
+    <title>Startsida</title>
+    <link rel="stylesheet" href="/style.css"/>
+    <script src="script.js" defer></script>
 </head>
 <body>
-    <h1>HALLÅ</h1>
+    <h1>ToDo-List </h1>
+    <div id="form1">
+    <form method="post" th:action="@{/}">
+        <input type="text" name="username" id="username" placeholder="Användarnamn">
+        <input type="text" name="password" id="password" placeholder="Lösenord">
+        <button id="loginBtn"> Logga in</button>
+    </form>
 </body>
 </html>
